requirements.txt
absl-py==2.1.0
accelerate==1.1.1
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitsandbytes==0.44.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
docstring_parser==0.16
durationpy==0.9
efficientnet_pytorch==0.7.1
einops==0.8.0
eval_type_backport==0.2.0
evaluate==0.4.3
eventlet==0.38.0
# faiss-cpu==1.9.0
PyQt5
fastapi==0.115.5
filelock==3.13.1
Flask==3.1.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==3.0.0
greenlet==3.1.1
grokfast-pytorch==0.0.10
grpcio==1.67.1
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hf_transfer==0.1.8
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httptools==0.6.4
httpx==0.13.3
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.12.0
kiwisolver==1.4.7
kubernetes==31.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.18
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==5.0.1
mne==1.8.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
ollama==0.3.3
onnxruntime==1.20.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
opt_einsum==3.4.0
optree==0.13.1
optuna==4.1.0
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peft==0.13.2
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
posthog==3.7.0
pretrainedmodels==0.7.4
propcache==0.2.0
protobuf==3.20.3
psutil==6.1.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytorch-triton-rocm==3.0.0+21eae954ef
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment-anything==1.0
segmentation-models-pytorch==0.3.4
sentence-transformers==3.3.0
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
simple-websocket==1.1.0
simsimd==6.0.7
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
split-folders==0.5.1
SQLAlchemy==2.0.35
starlette==0.41.2
stringzilla==3.10.10
supervision==0.25.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-rocm==2.13.1.600
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.9.7
tokenizers==0.20.3
torch==2.5.0.dev20240802+rocm6.0
torch-geometric==2.6.1
torchaudio==2.4.0.dev20240822+rocm6.0
torchmetrics==1.6.0
torchvision==0.20.0.dev20240822+rocm6.0
tqdm==4.67.0
transformers==4.46.2
triton==3.1.0
trl==0.12.1
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.9.0
tzdata==2024.2
tzlocal==5.2
unsloth==2024.11.7
unsloth_zoo==2024.11.5
urllib3==2.2.3
utils==1.0.2
uvicorn==0.32.0
uvloop==0.21.0
wandb==0.18.7
watchfiles==0.24.0
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wget==3.2
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
